requirements/requirements.txt
#
# This file is autogenerated by pip-compile with python 3.10
# To update, run:
#
#    pip-compile --output-file=requirements/requirements.txt requirements/requirements.in
#
asgiref==3.5.2
    # via django
async-timeout==4.0.2
    # via redis
boto3==1.23.1
    # via
    #   django-ses
    #   healthy-django
botocore==1.26.1
    # via
    #   boto3
    #   s3transfer
certifi==2021.10.8
    # via
    #   requests
    #   sentry-sdk
cffi==1.15.0
    # via cryptography
charset-normalizer==2.0.12
    # via requests
click==8.1.3
    # via pip-tools
colorama==0.4.4
    # via click
coreapi==2.3.3
    # via drf-yasg
coreschema==0.0.4
    # via
    #   coreapi
    #   drf-yasg
cryptography==36.0.2
    # via
    #   -r requirements/requirements.in
    #   django-ses
defusedxml==0.7.1
    # via odfpy
deprecated==1.2.13
    # via redis
diff-match-patch==20200713
    # via django-import-export
django==4.0.4
    # via
    #   -r requirements/requirements.in
    #   django-allow-cidr
    #   django-cors-headers
    #   django-filter
    #   django-import-export
    #   django-ses
    #   djangorestframework
    #   djangorestframework-simplejwt
    #   drf-yasg
django-allow-cidr==0.4.0
    # via -r requirements/requirements.in
django-cors-headers==3.12.0
    # via -r requirements/requirements.in
django-environ==0.8.1
    # via -r requirements/requirements.in
django-filter==21.1
    # via -r requirements/requirements.in
django-import-export==2.8.0
    # via -r requirements/requirements.in
django-ses[events]==3.0.1
    # via -r requirements/requirements.in
djangorestframework==3.13.1
    # via
    #   -r requirements/requirements.in
    #   djangorestframework-simplejwt
    #   drf-yasg
djangorestframework-simplejwt==5.1.0
    # via -r requirements/requirements.in
drf-yasg==1.20.0
    # via -r requirements/requirements.in
et-xmlfile==1.1.0
    # via openpyxl
gunicorn==20.1.0
    # via -r requirements/requirements.in
healthy-django==0.1.0
    # via -r requirements/requirements.in
idna==3.3
    # via requests
inflection==0.5.1
    # via drf-yasg
itypes==1.2.0
    # via coreapi
jinja2==3.1.2
    # via coreschema
jmespath==1.0.0
    # via
    #   boto3
    #   botocore
markdown==3.3.7
    # via -r requirements/requirements.in
markuppy==1.14
    # via tablib
markupsafe==2.1.1
    # via jinja2
netaddr==0.8.0
    # via django-allow-cidr
odfpy==1.4.1
    # via tablib
openpyxl==3.0.9
    # via tablib
packaging==21.3
    # via
    #   drf-yasg
    #   redis
pep517==0.12.0
    # via pip-tools
pip-tools==6.6.1
    # via -r requirements/requirements.in
psycopg2==2.9.3
    # via -r requirements/requirements.in
psycopg2-binary==2.9.3
    # via -r requirements/requirements.in
pycparser==2.21
    # via cffi
pyjwt==2.4.0
    # via djangorestframework-simplejwt
pyparsing==3.0.9
    # via packaging
python-dateutil==2.8.2
    # via botocore
pytz==2022.1
    # via
    #   django-ses
    #   djangorestframework
pyyaml==6.0
    # via tablib
redis==4.3.1
    # via -r requirements/requirements.in
requests==2.27.1
    # via
    #   coreapi
    #   django-ses
ruamel-yaml==0.17.21
    # via drf-yasg
ruamel-yaml-clib==0.2.6
    # via ruamel-yaml
s3transfer==0.5.2
    # via boto3
sentry-sdk==1.5.12
    # via -r requirements/requirements.in
six==1.16.0
    # via python-dateutil
sqlparse==0.4.2
    # via
    #   -r requirements/requirements.in
    #   django
tablib[html,ods,xls,xlsx,yaml]==3.2.1
    # via django-import-export
tomli==2.0.1
    # via pep517
tzdata==2022.1
    # via django
uritemplate==4.1.1
    # via
    #   coreapi
    #   drf-yasg
urllib3==1.26.9
    # via
    #   botocore
    #   requests
    #   sentry-sdk
wheel==0.37.1
    # via pip-tools
whitenoise==6.1.0
    # via -r requirements/requirements.in
wrapt==1.14.1
    # via deprecated
xlrd==2.0.1
    # via tablib
xlwt==1.3.0
    # via tablib

# The following packages are considered to be unsafe in a requirements file:
# pip
# setuptools
